Function calling w/small models fixed
There's two complex edge cases that interact:
1. Models such as Mistral 7B via Nous that don't have an
EOS token per se, but a set of tokens that add up to a
string that is the EOS token. This was covered in a previous
commit that maintains a buffer of tokens -> strings until
it is certain they aren't the EOS string.
2. StableLM Zephyr 3B's EOS token does not have a
string value from llama_token_to_piece. So, logic
that checks the buffer for the string version of the EOS
token, then flushes the buffer and breaks if present,
isn't triggered. This meant the loop would make it down to
the end, where it checks to see if the last token ID is the
EOS token ID, and breaks. But, it didn't flush the buffer,
leaving incomplete responses.