Create missingbuild-info-gen-cpp.cmake
The last commit worked consistently locally, but not on CI. CI failed with the same error that initiated the last commit: src/llama.cpp/common/cmake/build-info-gen-cpp.cmake missing. Now, the general CMakeLists.txt and the specific windows one both create a placefolder file here.

The local compile was succeeding and inferencing with an empty file there. It is also compiling, but was not tested for inferencing, with this placeholder file.