Wrap inference in try/catch
This was tested via a crash on macOS when a Q4_0_4_4 model is used with GPU enabled.

It is supposed to, in general, avoid crashes that bring down the host app due to issues in llama.cpp. This is not generally common, however, malformed GGUFs that can induce this behavior are not uncommon.