fix?: windows
the issue is that the select_best_resolution function is being exposed to C via the CLIP_API macro but returns a C++ type (std::pair). This isn't allowed since C has no concept of std::pair. The fix is to put this function in a C++ only section by wrapping it in #ifdef __cplusplus guards (since it's only used internally by other C++ code).
 