fix: accidentally removed adding input in prior commit
This was v interesting issue, took hours to track down because the example app and core app *seemingly* had taken the new dependency and worked just fine.

Seems builds were using the new llama.cpp changes but *not* the changes to the *fllama* code.

For the example app, it's very important to run flutter clean.