scaffold: log message previews before applying template
This provides useful info with silly templates like Gemma 3's, which
rejects the messages when they aren't purely alternating between
assistant and user.

This is hard to debug from Telosnex: ex. simply attaching an image
seems to be enough to trigger it on the Telosnex end. Yet, there's
no issue on the FLLAMA example app.