fix: cutoff logs
Modelling the logging callback as a call into Dart, that will be completely asynchronously at some point.

Retaining the string previously seemed to *help* with this issue, slightly, but it's visibly bad, especially with Gemma 3.

Maintaining a queue of the last N messages seems to help significantly with the log formatting.

FWIW, this was probably more visible with Gemma 3 because it has *1000s* of logs re: reversed tokens not marked as eog. These are rejected, but on the log *consumer* side.